Tratamento de erro - Alerta de conexão de internet

+++ MauiAppTempoAgora/Models/Tempo.cs
@@ -6,7 +6,7 @@
         public double? lat { get; set; }
         public double? temp_min { get; set; }
         public double? temp_max { get; set; }
-        public int? visibility { get; set; }
+        public double? visibility { get; set; }
         public double? speed { get; set; }
         public string? main { get; set; }
         public string? description { get; set; }
